730 move descr amount
